metering: check that we are talking to rrdached and not directly to file
on developer systems we are not talking through the rrdcached system
if by accident an rrdcached is running system wide, unit tests might
hang.

#0  __lll_lock_wait_private () at ../nptl/sysdeps/unix/sysv/linux/x86_64/lowlevellock.S:95
#1  0x00007f5bd191b530 in _L_lock_31 () from /lib/x86_64-linux-gnu/libc.so.6
#2  0x00007f5bd191b079 in do_system (line=0x8b440d "killall -9 rrdcached") at ../sysdeps/posix/system.c:68
#3  0x00000000005f9a4d in dss::Metering::checkDBReset (this=0x7ffceee3f5a0, _sampledAt=..., _rrdFileName='0' <repeats 33 times>, "d.rrd")
    at ../../build/../src/metering/metering.cpp:425
#4  0x00000000004d199e in testMetering::test_checkDBReset (this=0x7ffceee3f5a0, _sampledAt=..., _rrdFileName='0' <repeats 33 times>, "d.rrd")
    at ../../build/../tests/meteringtests.cpp:62

+++ src/metering/metering.cpp
@@ -421,8 +421,11 @@ static const int DISK_FLUSH_INTERVAL = 10*60; // ten minutes
       if (result < 0) {
         log(rrd_get_error(), lsError);
       }
-      if (system("killall -9 rrdcached") != 0) {
-        log("Restarting rrdcached was not successful.", lsWarning);
+      if (!m_RrdcachedPath.empty()) {
+        /* TODO sv restart rrdcached might hang */
+        if (system("killall -9 rrdcached") != 0) {
+          log("Restarting rrdcached was not successful.", lsWarning);
+        }
       }
     }
     return (result == 0);
